added custom hook for GameRow component
